*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3637 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func.c
@@ -61,7 +61,7 @@
 /* The following variables are shared with loader.c which adds further 
    functions from the loaded modules */
 
-int Num_Func;        /* number of built-in and listed functions */
+size_t Num_Func;     /* number of built-in and listed functions */
 Func *Fncts;         /* structure for list of functions */
 
 
@@ -219,7 +219,8 @@ static int func_cmp1( const void *a, const void *b )
 
 void functions_exit( void )
 {
-	int i;
+	size_t i;
+
 
 	if ( Fncts == NULL )
 		return;
@@ -339,7 +340,7 @@ Var *func_call( Var *f )
 	long abs_len;
 #ifndef NDEBUG
 	Func *cur_func;
-	int i;
+	size_t i;
 
 
 	/* Check (and double-check) that it's really a function variable - one
